refactor: SoptampPartRankCalculator를 이용하여 파트 랭킹을 계산할 수 있도록 개선 (#332)
