+++ src/main/java/org/sopt/app/application/rank/SoptampPartRankCalculator.java
@@ -1,44 +1,24 @@
 package org.sopt.app.application.rank;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.sopt.app.application.soptamp.SoptampPointInfo.PartRank;
 import org.sopt.app.common.exception.InternalServerException;
 import org.sopt.app.domain.enums.Part;
-import org.sopt.app.interfaces.postgres.soptamp_point.SoptampPointRepository;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 
-@Component
+
 @RequiredArgsConstructor
 public class SoptampPartRankCalculator {
 
-    private final SoptampPointRepository soptampPointRepository;
-
-    @Value("${sopt.current.generation}")
-    private Long currentGeneration;
-
-    private Map<Part, Long> partPoints;
+    private final Map<Part, Long> partPoints;
 
     public List<PartRank> findAllPartRanks() {
-        this.partPoints = findSumOfPointAllParts();
 
         return partPoints.keySet().stream()
                 .map(this::createPartRank)
                 .toList();
-    }
-
-    private Map<Part, Long> findSumOfPointAllParts() {
-        List<Part> allParts = Arrays.asList(Part.class.getEnumConstants());
-        return allParts.stream()
-                .collect(Collectors.toMap(
-                        part -> part,
-                        part -> soptampPointRepository.findSumOfPointBySamePartAndGeneration(part, currentGeneration))
-                );
     }
 
     private PartRank createPartRank(Part targetPart) {
